Was procrastinating so made perception side more readable
Better names, better function breakdowns, better comments
