+++ src/perception/yolo_pipeline/package.xml
@@ -3,7 +3,7 @@
 <package format="3">
     <name>yolo_pipeline</name>
     <version>0.0.0</version>
-    <description>YOLOv5 image processing in python</description>
+    <description>YOLOv5 image analysis in python</description>
     <maintainer email="qutmotorsport.team@gmail.com">QUT Motorsport</maintainer>
     <license>MIT</license>
