+++ src/perception/yolo_pipeline/setup.py
@@ -1,6 +1,7 @@
 from setuptools import setup
 
 package_name = 'yolo_pipeline'
+submodule_name = 'yolo_pipeline/sub_module'
 
 setup(
     name=package_name,
@@ -15,12 +16,11 @@ setup(
     zip_safe=True,
     maintainer='QUT Motorsport',
     maintainer_email='qutmotorsport.team@gmail.com',
-    description='A package to detect cones from a topic using yolov5',
+    description='YOLOv5 image analysis in python',
     license='MIT',
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
             'yolo = yolo_pipeline.yolo:main',
-            'test = yolo_pipeline.yolo_tester:main'
         ],
     },)
